Docker-based snapshot mechanism (PRv2) (#12)
* Init basic docker snapshotting

* New and improved docker-based snapshotting mechanism.

* added several "standalone" examples that use the docker-based snapshoting mechanism
* updated snapshot scripts for musl compatibility

* Add libfuzzer helper functions directly to the snapchange crate.
They should be generic enough to cover all fuzzers that use `LLVMFuzzerTestOneInput`.

* Added handling of run the fuzzer on an arbitrary division of the cores using `-c /N`.
Useful to launch on half of the cores (i.e., those that are non-hyperthreaded) with `-c /2` or on a third of the cores with `-c /3`, etc..

* Repository and script cleanup

* Removed duplicated scripts such as gdbsnapshot.py, coverage scripts, etc.
* Docker scripts do not hardcode `/snapchange/` anymore and instead rely
  on a `SNAPCHANGE_ROOT` env var. This allows to run the scripts outside
  of the docker container, if needed. This required also a slight change
  in paths of the scripts.
* All coverage scripts now reside in `docker/coverage_scripts` with a
  symlink from `utils/coverage_scripts`.
* Switched to ubuntu jammy aka. 22.04 for the snapchange docker
  container
* install rizin from source in the docker container instead of radare2
  via package manager.

* Two new scripts for extracting coverage breakpoints
One utilizes angr and the other uses rzpipe. Currently they are unused, but might be interesting in the future.

* updated src/_docs to include the new docker-based mechanism + fixed the fuzzer_template

* fixes for the ghidra coverage script

* Angr-based coverage script is now capable of generating a dictionary for snapchange fuzzers.
The script searches for integer constants and string constants used in comparisons.

There is a basic vex-based mode that is quite fast, but also a bit
limited. There is also a super-slow decompiler based search that tries
to decompile every function before searching for constants... ¯\_(ツ)_/¯
(disabled by default)

Maybe one can do limited under-constrained symex of whole functions to
get somewhat better results at better performance?

---------

Co-authored-by: Cory Duplantis <corydu@amazon.com>
Co-authored-by: Michael Rodler <mrodler@amazon.de>