support for rough (function-level) traces using coverage + symbol breakpoints (#25)
* Consistent use of `SymbolList` across the code.
`SymbolList` is now an alias for `Vec<Symbol>`, which can be binary searched. Requires the Vec to be sorted, but we just need to do this at construction time.

* add InputWithMetadata to prelude

* Support for tracing using breakpoints instead of single stepping
Using coverage and symbol breakpoints to obtain a trace. This allows to obtain a rough trace more quickly.

* Ignore integer overflows in `VirtAddr::offset` also in debug builds.

* set symbol breakpoints only on executable/non-writable memory.
otherwise setting breakpoints could break program logic (i.e., writing 0xff to some global int variable).

---------

Co-authored-by: Michael Rodler <mrodler@amazon.de>